game/engine/seongsu_rpg_project/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiosqlite
pydantic>=2.0
osmnx
matplotlib
requests
python-dotenv
